+++ maxent_gan/models/__init__.py
@@ -16,3 +16,4 @@ from .mlp import MLPDiscriminator, MLPGenerator  # noqa: F401
 from .sngan import SN_DCGAN_Generator  # noqa: F401
 from .sngan import SN_DCGAN_Discriminator, SN_ResNet_Generator32
 from .wgan import WGANDiscriminator  # noqa: F401
+from .wgan import WGANDiscriminatorIN, WGANGeneratorIN  # noqa: F401