+++ maxent_gan/models/wgan/__init__.py
@@ -1,1 +1,2 @@
 from .wgan_gp import WGANDiscriminator  # noqa: F401
+from .wgan_gp_in import WGANDiscriminatorIN, WGANGeneratorIN  # noqa: F401